move send notif function to utils.js
