+++ routes/utils.js
@@ -2,6 +2,9 @@ var Promise = require('promise');
 const winston = require('winston');
 const logger = winston.loggers.get('gomngr');
 var CONFIG = require('config');
+var GENERAL_CONFIG = CONFIG.general;
+const MAILER = CONFIG.general.mailer;
+var notif = require('../routes/notif_'+MAILER+'.js');
 
 const MongoClient = require('mongodb').MongoClient;
 var mongodb = null;
@@ -430,3 +433,72 @@ function _freeId (objType, id) {
         }
     });
 }
+
+
+
+// generic function for send notif (and email)
+
+function get_mail_config () {
+    var MAIL_CONFIG = {};
+    // todo: more and more ugly init...
+    if (CONFIG[MAILER]) { MAIL_CONFIG = CONFIG[MAILER]; }
+    // todo: find a cleaner way to allow registration if no mail are configured
+    if (!MAIL_CONFIG.origin) { MAIL_CONFIG.origin = 'nomail@nomail.org'; }
+    return MAIL_CONFIG;
+}
+
+async function gen_mail_opt (options, variables)
+{
+    var MAIL_CONFIG = get_mail_config();
+    // todo: check if each option exist and use default value
+    let name = options['name'];
+    let destinations = options['destinations'];
+    let subject = GENERAL_CONFIG.name + ' ' + options['subject'];
+
+    //find message
+    let message = CONFIG.message[name].join('\n');
+    let html_message = CONFIG.message[name + '_html'].join('');
+
+    // replace variable in message
+    for (let key in variables) {
+        let value = variables[key];
+        let re = new RegExp(key,"g");
+        message = message.replace(re, value);
+        html_message = html_message.replace(re, value);
+    };
+
+    // always add footer
+    message = message + '\n' + CONFIG.message.footer.join('\n');
+    html_message = html_message + '<br/>' + CONFIG.message.footer_html.join('<br/>');
+
+    // set mailOptions
+    let mailOptions = {
+        origin: MAIL_CONFIG.origin, // sender address
+        destinations:  destinations, // list of receivers
+        subject: subject, // Subject line
+        message: message, // plaintext body
+        html_message: html_message // html body
+    };
+
+    // tmp for debug
+    // logger.info(mailOptions);
+
+    // todo: find if we should return or send mail ...
+    return mailOptions;
+}
+
+async function send_notif_mail (options, variables) {
+    if(notif.mailSet()) {
+        try {
+            let mailOptions = await gen_mail_opt(options, variables);
+            await notif.sendUser(mailOptions);
+        } catch(err) {
+            logger.error('send notif mail error', err);
+        }
+    }
+}
+
+// exports mail functions
+exports.get_mail_config = get_mail_config;
+exports.gen_mail_opt = gen_mail_opt;
+exports.send_notif_mail = send_notif_mail;
